net: dhcpv4: moved event_callback to prevent deadlocks
To prevent deadlocks and incorrect behavior with multiple ethernet interfaces, the init_event_callback is moved to the net_dhcpv4_init function. 
Both add and remove event callback in the _start and _stop function have benn removed as they are not interface specific and creates a deadlock when using in the Carrier_on (link detect) event callback. 

Signed-off-by: Vincent van der Locht <vincent@vlotech.nl>