Added optional ETH phy thread for STM32 for link check
The ethernet driver of stm32 (low level driver) does not generate link information (cable attached or not). To allow applications to react on the link detect, an additional thread is added which poll the link status at a low rate.
This can be switched on/off through the menuconfig.
Implementation is aligned to mcux driver which has a irq for link check

Signed-off-by: Vincent van der Locht <vincent@vlotech.nl>